Fix fonttbl problem.
Create a better solution for the rtf parser arbitrarily adding font
names to the output. These appear to be in a fonttbl section towards the
beginning of the rtf, so rather than specifying the font names we want
to parse out, we parse out the entire fonttbl. We need txt only anyway,
and don't care about fonts.

Signed-off-by: Matthias Nott <matthias.nott@sap.com>